create function for eventListener click
